network: Move `ethtool` handlers to the driver package
The `driver` package is a better fit for the ethtool helpers, as it
already includes netlink and sysctl wrappers.
