+++ pkg/network/driver/BUILD.bazel
@@ -4,6 +4,7 @@ go_library(
     name = "go_default_library",
     srcs = [
         "common.go",
+        "ethtool.go",
         "generated_mock_common.go",
     ],
     importpath = "kubevirt.io/kubevirt/pkg/network/driver",
@@ -13,7 +14,6 @@ go_library(
         "//pkg/network/dhcp/server:go_default_library",
         "//pkg/network/dhcp/serverv6:go_default_library",
         "//pkg/network/dns:go_default_library",
-        "//pkg/network/link:go_default_library",
         "//pkg/util/sysctl:go_default_library",
         "//pkg/virt-handler/selinux:go_default_library",
         "//pkg/virt-launcher/virtwrap/converter:go_default_library",
