+++ pkg/network/link/BUILD.bazel
@@ -4,7 +4,6 @@ go_library(
     name = "go_default_library",
     srcs = [
         "address.go",
-        "ethtool.go",
         "names.go",
         "reserved_macs.go",
     ],
